nest support request controller
