+++ app/views/lockbox_partners/show.html.erb
@@ -28,7 +28,7 @@
               <td><%= action.amount %></td>
               <td>
                 <% if action.support_request %>
-                  <%= link_to support_request_path(action.support_request) do %>
+                  <%= link_to lockbox_partner_support_request_path(action.support_request) do %>
                     View details <i class="fa fa-long-arrow-right" aria-hidden="true"></i>
                   <% end %>
                 <% end %>
