+++ config/routes.rb
@@ -5,13 +5,14 @@ Rails.application.routes.draw do
     registrations: 'users/registrations'
   }
 
-  resources :support_requests, only: [:new, :create, :show]
+  resources :support_requests, only: [:new, :create]
 
   root to: 'dashboard#index'
 
   resources :lockbox_partners, only: [:new, :create, :show] do
     scope module: 'lockbox_partners' do
       resources :users, only: [:new, :create]
+      resources :support_requests, only: [:new, :create, :show]
       resource :add_cash, only: [:new, :create], controller: 'add_cash'
     end
   end
